+++ Src/SOS.DataStewardship.Api/Endpoints/Events/GetEventByIdEndpoint.cs
@@ -3,33 +3,33 @@ using SOS.DataStewardship.Api.Extensions;
 using SOS.DataStewardship.Api.Application.Managers.Interfaces;
 using SOS.DataStewardship.Api.Contracts.Enums;
 
-namespace SOS.DataStewardship.Api.Endpoints.DataStewardship;
+namespace SOS.DataStewardship.Api.Endpoints.Events;
 
 public class GetEventByIdEndpoint : IEndpointDefinition
 {
     public void DefineEndpoint(WebApplication app)
     {
-        app.MapGet("/datastewardship/events/{id}", GetEventByIdAsync)
+        app.MapGet("/events/{id}", GetEventByIdAsync)
             .Produces<Contracts.Models.Event>(StatusCodes.Status200OK, "application/json")
             .Produces<ProblemDetails>(StatusCodes.Status404NotFound)
             .Produces<HttpValidationProblemDetails>(StatusCodes.Status400BadRequest)
             .Produces<ProblemDetails>(StatusCodes.Status500InternalServerError);
     }
-    
+
     [SwaggerOperation(
         Description = "Get event by Id. Example: urn:lsid:swedishlifewatch.se:dataprovider:Artportalen:event:10002293427000658739",
         OperationId = "GetEventById",
         Tags = new[] { "DataStewardship" })]
-    [SwaggerResponse(404, "Not Found - The requested event doesn't exist")]    
+    [SwaggerResponse(404, "Not Found - The requested event doesn't exist")]
     private async Task<IResult> GetEventByIdAsync(IDataStewardshipManager dataStewardshipManager,
         [FromRoute, SwaggerParameter("The event id", Required = true)] string id,
         [FromQuery, SwaggerParameter("The export mode")] ExportMode exportMode = ExportMode.Json,
         [FromQuery, SwaggerParameter("The response coordinate system")] CoordinateSystem responseCoordinateSystem = CoordinateSystem.EPSG4326)
-    {        
+    {
         var eventModel = await dataStewardshipManager.GetEventByIdAsync(id, responseCoordinateSystem);
         if (eventModel == null) return NotFoundResult(id);
 
-        return exportMode.Equals(ExportMode.Csv) ? Results.File(eventModel.ToCsv(), "text/tab-separated-values", "dataset.csv") : Results.Ok(eventModel);       
+        return exportMode.Equals(ExportMode.Csv) ? Results.File(eventModel.ToCsv(), "text/tab-separated-values", "dataset.csv") : Results.Ok(eventModel);
     }
 
     private IResult NotFoundResult(string id)