+++ Tests/Integration/SOS.DataStewardship.Api.IntegrationTests/Tests/Events/Search/EventsEventIdsFilterTests.cs
@@ -25,7 +25,7 @@ public class EventsEventIdsFilterTests : TestBase
 
         // Act
         var pageResult = await ApiClient.PostAndReturnAsJsonAsync<PagedResult<Event>, EventsFilter>(
-            $"datastewardship/events?skip=0&take=0", searchFilter, jsonSerializerOptions);
+            $"events?skip=0&take=0", searchFilter, jsonSerializerOptions);
 
         // Assert
         pageResult.TotalCount.Should().Be(1);