+++ Tests/Integration/SOS.DataStewardship.Api.IntegrationTests/Tests/Events/Search/EventsGeographicsFilterTests.cs
@@ -30,7 +30,7 @@ public class EventsGeographicsFilterTests : TestBase
 
         // Act
         var pageResult = await ApiClient.PostAndReturnAsJsonAsync<PagedResult<Event>, EventsFilter>(
-            $"datastewardship/events?skip=0&take=0", searchFilter, jsonSerializerOptions);
+            $"events?skip=0&take=0", searchFilter, jsonSerializerOptions);
 
         // Assert
         pageResult.TotalCount.Should().Be(1);
@@ -55,7 +55,7 @@ public class EventsGeographicsFilterTests : TestBase
 
         // Act
         var pageResult = await ApiClient.PostAndReturnAsJsonAsync<PagedResult<Event>, EventsFilter>(
-            $"datastewardship/events?skip=0&take=0", searchFilter, jsonSerializerOptions);
+            $"events?skip=0&take=0", searchFilter, jsonSerializerOptions);
 
         // Assert
         pageResult.TotalCount.Should().Be(1);
@@ -80,7 +80,7 @@ public class EventsGeographicsFilterTests : TestBase
 
         // Act
         var pageResult = await ApiClient.PostAndReturnAsJsonAsync<PagedResult<Event>, EventsFilter>(
-            $"datastewardship/events?skip=0&take=0", searchFilter, jsonSerializerOptions);
+            $"events?skip=0&take=0", searchFilter, jsonSerializerOptions);
 
         // Assert
         pageResult.TotalCount.Should().Be(1);
@@ -105,7 +105,7 @@ public class EventsGeographicsFilterTests : TestBase
 
         // Act
         var pageResult = await ApiClient.PostAndReturnAsJsonAsync<PagedResult<Event>, EventsFilter>(
-            $"datastewardship/events?skip=0&take=0", searchFilter, jsonSerializerOptions);
+            $"events?skip=0&take=0", searchFilter, jsonSerializerOptions);
 
         // Assert
         pageResult.TotalCount.Should().Be(1);
@@ -130,7 +130,7 @@ public class EventsGeographicsFilterTests : TestBase
 
         // Act
         var pageResult = await ApiClient.PostAndReturnAsJsonAsync<PagedResult<Event>, EventsFilter>(
-            $"datastewardship/events?skip=0&take=0", searchFilter, jsonSerializerOptions);
+            $"events?skip=0&take=0", searchFilter, jsonSerializerOptions);
 
         // Assert
         pageResult.TotalCount.Should().Be(1);
@@ -155,7 +155,7 @@ public class EventsGeographicsFilterTests : TestBase
 
         // Act
         var pageResult = await ApiClient.PostAndReturnAsJsonAsync<PagedResult<Event>, EventsFilter>(
-            $"datastewardship/events?skip=0&take=0", searchFilter, jsonSerializerOptions);
+            $"events?skip=0&take=0", searchFilter, jsonSerializerOptions);
 
         // Assert
         pageResult.TotalCount.Should().Be(1);