+++ Tests/Integration/SOS.DataStewardship.Api.IntegrationTests/Tests/Occurrences/Search/OccurrencesDateFilterTests.cs
@@ -37,7 +37,7 @@ public class OccurrencesDateFilterTests : TestBase
         
         // Act
         var pageResult = await ApiClient.PostAndReturnAsJsonAsync<PagedResult<Occurrence>, OccurrenceFilter>(
-            $"datastewardship/occurrences?skip=0&take=0", searchFilter, jsonSerializerOptions);
+            $"occurrences?skip=0&take=0", searchFilter, jsonSerializerOptions);
 
         // Assert
         pageResult.TotalCount.Should().Be(6);
@@ -69,7 +69,7 @@ public class OccurrencesDateFilterTests : TestBase
 
         // Act
         var pageResult = await ApiClient.PostAndReturnAsJsonAsync<PagedResult<Occurrence>, OccurrenceFilter>(
-            $"datastewardship/occurrences?skip=0&take=0", searchFilter, jsonSerializerOptions);
+            $"occurrences?skip=0&take=0", searchFilter, jsonSerializerOptions);
 
         // Assert
         pageResult.TotalCount.Should().Be(3);
@@ -98,7 +98,7 @@ public class OccurrencesDateFilterTests : TestBase
 
         // Act
         var pageResult = await ApiClient.PostAndReturnAsJsonAsync<PagedResult<Occurrence>, OccurrenceFilter>(
-            $"datastewardship/occurrences?skip=0&take=0", searchFilter, jsonSerializerOptions);
+            $"occurrences?skip=0&take=0", searchFilter, jsonSerializerOptions);
 
         // Assert
         pageResult.TotalCount.Should().Be(2);
@@ -127,7 +127,7 @@ public class OccurrencesDateFilterTests : TestBase
 
         // Act
         var pageResult = await ApiClient.PostAndReturnAsJsonAsync<PagedResult<Occurrence>, OccurrenceFilter>(
-            $"datastewardship/occurrences?skip=0&take=0", searchFilter, jsonSerializerOptions);
+            $"occurrences?skip=0&take=0", searchFilter, jsonSerializerOptions);
 
         // Assert
         pageResult.TotalCount.Should().Be(2);