+++ Tests/Integration/SOS.DataStewardship.Api.IntegrationTests/Tests/Occurrences/Search/OccurrencesTaxonFilterTests.cs
@@ -24,7 +24,7 @@ public class OccurrencesTaxonFilterTests : TestBase
         
         // Act
         var pageResult = await ApiClient.PostAndReturnAsJsonAsync<PagedResult<Occurrence>, OccurrenceFilter>(
-            $"datastewardship/occurrences?skip=0&take=0", searchFilter, jsonSerializerOptions);
+            $"occurrences?skip=0&take=0", searchFilter, jsonSerializerOptions);
 
         // Assert
         pageResult.TotalCount.Should().Be(1);