+++ Docs/DataStewardship/CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog - Nature data - systematic species observations
 
 ## Version 1.0.5 (Unreleased)
-- Internal refactoring.
+
+### Fixed
+- Change urls from `/datastewardship/occurrences` to `/occurrences`.
 
 ## Version 1.0.4 (2023-06-12)
 - This is the first public release of the API.